requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
docopt==0.6.2
dotmap==1.3.30
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
Flask-Admin==1.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mjml==0.11.0
mypy==1.10.0
mypy-extensions==1.0.0
orjson==3.10.3
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
PyMySQL==1.1.1
pystache==0.6.5
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlalchemy-stubs==0.4
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
uvicorn==0.30.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
WTForms==3.1.2
